Working on migration to single nicer redux store
