+++ app/package.json
@@ -126,6 +126,7 @@
         "react-router": "^6.26.0",
         "react-router-dom": "^6.26.0",
         "react-router-hash-link": "^2.4.3",
+        "redux-persist": "^6.0.0",
         "rollup-plugin-node-polyfills": "^0.2.1",
         "stream": "^0.0.2",
         "typescript": "^5.5.4",
