--- app/src/context/authStore.ts
@@ -1,25 +1,0 @@
-import {configureStore} from '@reduxjs/toolkit';
-import {TypedUseSelectorHook, useDispatch, useSelector} from 'react-redux';
-
-// Import your reducers
-import authReducer from './slices/authSlice';
-
-// Configure the store
-export const store = configureStore({
-  reducer: {
-    auth: authReducer,
-  },
-  middleware: getDefaultMiddleware =>
-    getDefaultMiddleware({
-      serializableCheck: {},
-    }),
-  devTools: process.env.NODE_ENV !== 'production',
-});
-
-// Infer types from store
-export type RootState = ReturnType<typeof store.getState>;
-export type AppDispatch = typeof store.dispatch;
-
-// Create typed hooks
-export const useAppDispatch: () => AppDispatch = useDispatch;
-export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
